🎨 pageの構成を変更
