+++ src/app/(pages)/chart/page.tsx
@@ -1,0 +1,11 @@
+import { MyChart } from "@/app/components/MyChart";
+
+export default function Page() {
+  return (
+    <div className="flex justify-center">
+      <div className="w-1/2">
+        <MyChart />
+      </div>
+    </div>
+  );
+}
